+++ fh_cache.h
@@ -18,6 +18,6 @@ char *fh_decomp(nfs_fh3 fh);
 unfs3_fh_t fh_comp(const char *path, struct svc_req *rqstp, int need_dir);
 unfs3_fh_t *fh_comp_ptr(const char *path, struct svc_req *rqstp, int need_dir);
 
-char *fh_cache_add(uint32 dev, uint32 ino, const char *path);
+char *fh_cache_add(uint32 dev, uint64 ino, const char *path);
 
 #endif
